wip: make bookable status on items, not items types
also add many settings related to booking

+++ src/models/Scheduler.php
@@ -77,9 +77,11 @@ class Scheduler implements RestInterface
 
         $start = $this->normalizeDate($reqBody['start']);
         $end = $this->normalizeDate($reqBody['end'], true);
+        $this->checkOverlap($start, $end);
+        $this->checkSlotTime($start, $end);
 
         // users won't be able to create an entry in the past
-        $this->isFutureOrExplode(DateTime::createFromFormat(DateTime::ISO8601, $start));
+        $this->isFutureOrExplode(DateTime::createFromFormat(DateTime::ATOM, $start));
 
         // fix booking at midnight on monday not working. See #2765
         // we add a second so it works
@@ -177,6 +179,19 @@ class Scheduler implements RestInterface
     public function destroy(): bool
     {
         $this->canWriteOrExplode();
+        $event = $this->readOne();
+        if ($event['book_is_cancellable'] === 0 && !$this->Items->Users->isAdmin) {
+            throw new ImproperActionException(_('Event cancellation is not permitted.'));
+        }
+        if ($event['book_cancel_minutes'] !== 0 && !$this->Items->Users->isAdmin) {
+            $now = new DateTimeImmutable();
+            $eventStart = new DateTimeImmutable($event['start']);
+            $interval = $now->diff($eventStart);
+            $totalMinutes = ($interval->h * 60) + $interval->i;
+            if ($totalMinutes < $event['book_cancel_minutes']) {
+                throw new ImproperActionException(sprintf(_('Cannot cancel slot less than %d minutes before its start.'), $event['book_cancel_minutes']));
+            }
+        }
         $sql = 'DELETE FROM team_events WHERE id = :id';
         $req = $this->Db->prepare($sql);
         $req->bindParam(':id', $this->id, PDO::PARAM_INT);
@@ -233,6 +248,9 @@ class Scheduler implements RestInterface
      */
     private function updateEpoch(string $column, string $epoch): bool
     {
+        $event = $this->readOne();
+        $this->checkOverlap($event['start'], $event['end']);
+        $this->checkSlotTime($event['start'], $event['end']);
         $new = DateTimeImmutable::createFromFormat('U', $epoch);
         if ($new === false) {
             throw new ImproperActionException('Invalid date format received.');
@@ -249,7 +267,7 @@ class Scheduler implements RestInterface
 
     private function readOneEvent(): array
     {
-        $sql = 'SELECT team_events.id, team_events.team, team_events.item, team_events.start, team_events.end, team_events.title, team_events.userid, team_events.experiment, team_events.item_link,
+        $sql = 'SELECT team_events.id, team_events.team, team_events.item, team_events.start, team_events.end, team_events.title, team_events.userid, team_events.experiment, team_events.item_link, items.book_is_cancellable, items.book_cancel_minutes,
             team_events.title AS title_only,
             experiments.title AS experiment_title,
             items_linkt.title AS item_link_title
@@ -261,7 +279,9 @@ class Scheduler implements RestInterface
         $req = $this->Db->prepare($sql);
         $req->bindParam(':id', $this->id, PDO::PARAM_INT);
         $this->Db->execute($req);
-        return $this->Db->fetch($req);
+        $event = $this->Db->fetch($req);
+        $this->Items->setId($event['item']);
+        return $event;
     }
 
     /**
@@ -272,8 +292,8 @@ class Scheduler implements RestInterface
     private function updateStart(array $delta): bool
     {
         $event = $this->readOne();
-        $oldStart = DateTime::createFromFormat(DateTime::ISO8601, $event['start']);
-        $oldEnd = DateTime::createFromFormat(DateTime::ISO8601, $event['end']);
+        $oldStart = DateTime::createFromFormat(DateTime::ATOM, $event['start']);
+        $oldEnd = DateTime::createFromFormat(DateTime::ATOM, $event['end']);
         $seconds = '0';
         if (strlen((string) $delta['milliseconds']) > 3) {
             $seconds = substr((string) $delta['milliseconds'], 0, -3);
@@ -282,6 +302,8 @@ class Scheduler implements RestInterface
         $this->isFutureOrExplode($newStart);
         $newEnd = $oldEnd->modify('+' . $delta['days'] . ' day')->modify('+' . $seconds . ' seconds'); // @phpstan-ignore-line
         $this->isFutureOrExplode($newEnd);
+        $this->checkOverlap($newStart->format(DateTime::ATOM), $newEnd->format(DateTime::ATOM));
+        $this->checkSlotTime($newStart->format(DateTime::ATOM), $newEnd->format(DateTime::ATOM));
 
         $sql = 'UPDATE team_events SET start = :start, end = :end WHERE team = :team AND id = :id';
         $req = $this->Db->prepare($sql);
@@ -300,13 +322,15 @@ class Scheduler implements RestInterface
     private function updateEnd(array $delta): bool
     {
         $event = $this->readOne();
-        $oldEnd = DateTime::createFromFormat(DateTime::ISO8601, $event['end']);
+        $oldEnd = DateTime::createFromFormat(DateTime::ATOM, $event['end']);
         $seconds = '0';
         if (strlen((string) $delta['milliseconds']) > 3) {
             $seconds = substr((string) $delta['milliseconds'], 0, -3);
         }
         $newEnd = $oldEnd->modify('+' . $delta['days'] . ' day')->modify('+' . $seconds . ' seconds'); // @phpstan-ignore-line
         $this->isFutureOrExplode($newEnd);
+        $this->checkOverlap($event['start'], $newEnd->format(DateTime::ATOM));
+        $this->checkSlotTime($event['start'], $newEnd->format(DateTime::ATOM));
 
         $sql = 'UPDATE team_events SET end = :end WHERE team = :team AND id = :id';
         $req = $this->Db->prepare($sql);
@@ -340,6 +364,46 @@ class Scheduler implements RestInterface
         return $this->Db->execute($req);
     }
 
+    private function checkSlotTime(string $start, string $end): void
+    {
+        if ($this->Items->entityData['book_max_minutes'] === 0) {
+            return;
+        }
+        $start = new DateTimeImmutable($start);
+        $end = new DateTimeImmutable($end);
+        $interval = $start->diff($end);
+        $totalMinutes = ($interval->h * 60) + $interval->i;
+        if ($totalMinutes > $this->Items->entityData['book_max_minutes']) {
+            throw new ImproperActionException(sprintf(_('Slot time is limited to %d minutes.'), $this->Items->entityData['book_max_minutes']));
+        }
+
+    }
+
+    /**
+     * Look if another slot is present for the same item at the same time and throw exception if yes
+     */
+    private function checkOverlap(string $start, string $end): void
+    {
+        if ($this->Items->entityData['book_can_overlap'] === 1) {
+            return;
+        }
+        $sql = 'SELECT id FROM team_events WHERE :start < end AND :end > start AND item = :item';
+        if ($this->id) {
+            $sql .= ' AND id != :id';
+        }
+        $req = $this->Db->prepare($sql);
+        $req->bindParam(':start', $start, PDO::PARAM_STR);
+        $req->bindParam(':end', $end, PDO::PARAM_STR);
+        $req->bindParam(':item', $this->Items->id, PDO::PARAM_INT);
+        if ($this->id) {
+            $req->bindParam(':id', $this->id, PDO::PARAM_INT);
+        }
+        $this->Db->execute($req);
+        if (!empty($req->fetchAll())) {
+            throw new ImproperActionException(_('Overlapping booking slots is not permitted.'));
+        }
+    }
+
     /**
      * Check that the date is in the future
      * Unlike Admins, Users can't create/modify something in the past
@@ -360,12 +424,12 @@ class Scheduler implements RestInterface
     }
 
     /**
-     * Date can be Y-m-d or ISO::8601
+     * Date can be Y-m-d or ISO::ATOM
      * Make sure we have the time, too
      */
     private function normalizeDate(string $date, bool $rmDay = false): string
     {
-        if (DateTime::createFromFormat(DateTime::ISO8601, $date) === false) {
+        if (DateTime::createFromFormat(DateTime::ATOM, $date) === false) {
             $dateOnly = DateTime::createFromFormat('Y-m-d', $date);
             if ($dateOnly === false) {
                 throw new ImproperActionException('Could not understand date format!');
@@ -375,7 +439,7 @@ class Scheduler implements RestInterface
             if ($rmDay) {
                 $dateOnly->modify('-3min');
             }
-            return $dateOnly->format(DateTime::ISO8601);
+            return $dateOnly->format(DateTime::ATOM);
         }
         return $date;
     }
